Add files via upload
In this project I changed the text in the hello world to say,  "Hello, world.  This is my first app!". 

Also in this project, I had to change the text to the button to say when clicked, "You have clicked the button"

This is my first time making an app on android studio, and getting used to everything will take some time.  But I'm excited to learn!